move config out from main.go
